fix(orc8r): Make check-in status check in the backend (#12699)
At the moment, gateway health in NMS is checked by comparing the
last check-in time from the backend to the current time in the
browser. This check is unstable and results in an unreliable
health status in many cases.

With this change, the check-in time check is instead performed
in the backend, taking the configured check-in interval into
account. A follow-up commit will use the result in NMS.

Signed-off-by: Sebastian Thomas <sebastian.thomas@tngtech.com>